backend/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
SQLAlchemy
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-multipart
geopy
requests
razorpay
